Move telemetry name mapping shared code to parametric utils.py
Avoids a cross-test-file import between test_dynamic_configuration.py
and test_telemetry.py, per PR review (CI optimization requirement).
